Rename RAMInterpreter to RamInterpreter
